Move bulk alt text tests into the main alt text file
